// LondonFhirService.Manage.Client/src/models/foundations/fhir/AllergyIntoleranceData.ts
export type AllergyIntoleranceData = {
    id: string;
    display: string | null;
    code: string | null;
    system: string | null;
    clinicalStatus: string | null;
    type: string | null;
    verificationStatus: string | null;
    onsetDateTime: string | null;
    asserterRef: string | null;
};